requirements.txt
Django==1.11.5
djangorestframework==3.5.3
django-extra-fields==0.9
django-rest-framework-mongoengine==3.3.1
Gunicorn==19.6.0
mongoengine==0.13.0
mysqlclient==1.3.10
MySQL-python==1.2.5
scrapy==1.4.0
cryptography==2.0
requests==2.18.4